refactor whisper
