+++ screenpipe-audio/src/lib.rs
@@ -9,6 +9,7 @@ mod tokenizer;
 pub mod vad_engine;
 pub mod whisper;
 
+pub use audio_processing::resample;
 pub use core::{
     default_input_device, default_output_device, get_device_and_config, list_audio_devices,
     parse_audio_device, record_and_transcribe, trigger_audio_permission, AudioDevice, AudioStream,
@@ -16,5 +17,5 @@ pub use core::{
 };
 pub use encode::encode_single_audio;
 pub use pcm_decode::pcm_decode;
-pub use stt::{create_whisper_channel, resample, stt, AudioInput, TranscriptionResult};
+pub use stt::{create_whisper_channel, stt, AudioInput, TranscriptionResult};
 pub use vad_engine::VadEngineEnum;
