+++ screenpipe-audio/src/whisper/mod.rs
@@ -1,0 +1,7 @@
+mod decoder;
+mod model;
+mod process_chunk;
+
+pub use decoder::*;
+pub use model::*;
+pub use process_chunk::*;
